+++ CHANGELOG.md
@@ -28,6 +28,6 @@ Initial public release of Torch-RGCN.
 - Added LICENSE.
 - Added AIFB, AM, BGS, FB15k, FB15k-237, MUTAG, WN18 and WN18RR to data.
 - Added unit tests under [tests](tests).
-- Added scripts for node classification and link prediction under [experiments](experiments).
+- Added scripts for node classification and relation Prediction  under [experiments](experiments).
 - Added hyper-parameters under [configs](configs).
 - Added RGCN module under [rgcn](rgcn).